refactor: Split out args
